Tema04 - Flask Autenticación/Clase08/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
psycopg2-binary
python-dotenv
